made apis

+++ controllers/api/v1/posts_api.js
@@ -1,5 +1,6 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment')
+
 module.exports.index = async function(req, res){
     let posts = await Post.find({})
         .sort('-createdAt')
@@ -17,19 +18,22 @@ module.exports.index = async function(req, res){
 }
 
 module.exports.delete = async function (req, res) {
+   
     let post = await Post.findById(req.params.postId);
+   console.log((post.user == req.user.id))
     try {
-        //if (post.user == req.user.id) {
-            let p = post;
+        if (post.user == req.user.id) {
+        var p = post;
+           console.log('try if')
             post.remove();
             await Comment.deleteMany({ post: req.params.postId });
-            
-
-
-            
-        // }else{
-        //     return res.redirect('back')
-        // }
+        }
+        else{
+            console.log('try else')
+            return res.status(500).json({
+                message: "not authorized to delete"
+            })
+        }
         return res.status(200).json({
             message: "This post and comments",
             post: p,
@@ -37,7 +41,7 @@ module.exports.delete = async function (req, res) {
     } catch {
 
         return res.status(500).json({
-            message: "Error occured while deleting post",
+            message: "Internal server error",
         });
     }
 }
